+++ examples/PID_PonM/PID_PonM.ino
@@ -9,7 +9,7 @@
 #include <PID_v1.h>
 
 //Define Variables we'll be connecting to
-double Setpoint, Input, Output;
+float Setpoint, Input, Output;
 
 //Specify the links and initial tuning parameters
 PID myPID(&Input, &Output, &Setpoint, 2, 5, 1, P_ON_M, DIRECT);    //P_ON_M specifies that Proportional on Measurement be used
